exaspoon/exaspoon2/pyproject.toml
[project]
name = "exaspoon2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "supabase",
    "fastmcp",
    "httpx",
    "fastapi",
    "uvicorn[standard]",
    "websockets",
    "python-dotenv>=1.2.1",
    "spoon-ai-sdk"
]

[tool.uv.sources]
spoon-ai-sdk = { path = "./spoon-core" }
